requirements.txt
py-cord
python-dotenv
PyNaCl
lavalink
spotipy